requirements.txt
selenium>=4.0.0
webdriver-manager==4.0.1
requests==2.31.0
tqdm==4.66.2
python-dotenv>=0.19.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
beautifulsoup4>=4.9.3
click>=8.0.0
pre-commit>=3.3.3
pytest>=7.3.1
pytest-cov>=4.1.0
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
flake8-docstrings>=1.7.0
